Added PMIN/PMAX/PMUL/PCMP/PSHUF/PUNPCK/UNPCK/PACK instructions to the new emitter [this basically finishes all MMX instructions -- just some SSE2/SSE3 and SSE4 mess left!]
git-svn-id: http://pcsx2.googlecode.com/svn/trunk@1035 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ pcsx2/x86/ix86/implement/group1.h
@@ -134,7 +134,7 @@ public:
 // This class combines x86 with SSE/SSE2 logic operations (ADD, OR, and NOT).
 // Note: ANDN [AndNot] is handled below separately.
 //
-template< G1Type InstType, u8 OpcodeSSE >
+template< G1Type InstType, u16 OpcodeSSE >
 class xImpl_G1Logic : public xImpl_Group1<InstType>
 {
 public:
@@ -149,7 +149,7 @@ public:
 // ------------------------------------------------------------------------
 // This class combines x86 with SSE/SSE2 arithmetic operations (ADD/SUB).
 //
-template< G1Type InstType, u8 OpcodeSSE >
+template< G1Type InstType, u16 OpcodeSSE >
 class xImpl_G1Arith : public xImpl_G1Logic<InstType, OpcodeSSE >
 {
 public:
